Character Creator Save Functionality
Character Creator now has a button to save the character. Character saves are ignored using a .gitignore file so there are no conflicts because we have different characters. There is a test scene to test the "load character" function and a script that should be able to be reused in any game to load the character.